+++ docs/codex/recycle-sell/03-refactor/S4_스토어분리_설계및구현보고서.md
@@ -1,0 +1,93 @@
+﻿# S4 스토어 분리 리팩토링 설계 및 구현 보고서
+
+## 1. 작업 개요
+- 목표: 단일 `circularStock.js`에 혼재된 판매/재고/등록 draft/추천/ESG 책임을 분리해 유지보수성과 안정성을 높인다.
+- 적용 방식: 전면 전환(뷰/호출부를 새 스토어로 직접 교체), 기존 단일 스토어 제거.
+- 원칙: 기능 회귀 없이 책임 경계 고정, 기존 액션명 최대 유지.
+
+---
+
+## 2. 설계 요약
+## 2.1 책임 분리
+- Sale Store(`useCircularStockSaleStore`)
+  - 판매 목록/상세 조회
+  - 판매 등록 draft 상태 및 검증/제출
+  - AI 추천/선택 관련 상태
+  - ESG 스냅샷 계산
+- Inventory Store(`useCircularStockInventoryStore`)
+  - 재고 목록 조회/필터/정렬/페이징 상태
+  - 재고 로딩 액션
+  - 재고 row 매핑 및 조회(`getInventoryById`)
+- Common 유틸(`circularStockCommon.js`)
+  - 반올림/소재명 정규화/소재 판정/로컬스토리지 유틸 등 공통 함수
+
+## 2.2 의존 방향
+- `Sale -> Inventory` 단방향 참조 허용
+  - draft 등록 시 재고 검증에 `inventoryStore.getInventoryById` 사용
+- `Inventory -> Sale` 참조 금지
+- `auth.logout -> Sale.clearDraft`로 초기화 경로 고정
+
+---
+
+## 3. 구현 내역
+## 3.1 스토어
+- 추가
+  - `src/stores/hq/circularStock/circularStockSale.js`
+  - `src/stores/hq/circularStock/circularStockInventory.js`
+  - `src/stores/hq/circularStock/circularStockCommon.js`
+- 제거
+  - `src/stores/hq/circularStock/circularStock.js`
+
+## 3.2 호출부 전환
+- `useCircularStockStore` 제거 후 아래로 전환
+  - 판매 화면군: `useCircularStockSaleStore`
+  - 재고 관련 조회/필터: `useCircularStockInventoryStore`
+- 적용 파일
+  - `src/views/hq/circular-stock/HqCircularStockSalesRegisterView.vue`
+  - `src/views/hq/circular-stock/HqCircularStockSalesSkuSelectView.vue`
+  - `src/views/hq/circular-stock/HqCircularStockSalesHistoryView.vue`
+  - `src/views/hq/circular-stock/HqCircularStockSalesDetailView.vue`
+  - `src/views/hq/circular-stock/HqCircularStockInventoryView.vue`
+  - `src/stores/auth.js`
+
+## 3.3 로그아웃 초기화
+- `auth.logout`에서 `useCircularStockSaleStore().clearDraft()` 호출하도록 변경.
+- 효과: 로그아웃 후 판매 등록 draft/step/recommendation 상태 잔존 방지.
+
+---
+
+## 4. 인터페이스 변경
+- 제거: `useCircularStockStore`
+- 추가:
+  - `useCircularStockSaleStore`
+  - `useCircularStockInventoryStore`
+- 유지된 핵심 액션명(마이그레이션 리스크 완화)
+  - `fetchCircularSalesPage`
+  - `fetchCircularSaleDetail`
+  - `submitCircularStockSale`
+  - `loadCircularInventoryRows`
+  - `clearDraft`
+
+---
+
+## 5. 검증 결과
+- FE 빌드
+  - 명령: `npm.cmd run build`
+  - 결과: 성공
+- FE lint
+  - 명령: `npm.cmd run lint`
+  - 결과: 스크립트 부재로 실행 불가 (`Missing script: "lint"`)
+
+---
+
+## 6. 리스크 및 후속 권장
+- 일부 뷰 파일은 과거 인코딩/태그 깨짐 이력이 있어, 후속으로 문구/인코딩 일괄 점검 권장.
+- 이번 단계는 구조 분리에 집중했으며, 도메인 정책 변경(상태 정책 등)은 포함하지 않음.
+- 다음 단계에서 가능한 개선
+  - Sale Store 내부 계산 로직(ESG/추천 계산) 추가 모듈 분리
+  - 단위 테스트(스토어 액션/매핑/검증) 보강
+
+---
+
+## 7. 결론
+S4 목표대로 단일 `circularStock` 의존을 제거하고 판매/재고 스토어를 전면 분리했다. 책임 경계가 명확해져 화면별 의존성이 단순해졌고, 로그아웃 초안 초기화 경로도 Sale Store로 고정되어 상태 일관성이 개선되었다.
