+++ src/views/hq/circular-stock/HqCircularStockSalesDetailView.vue
@@ -1,14 +1,14 @@
-<script setup>
+﻿<script setup>
 import { computed, onMounted, ref } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
 import AppLayout from '@/components/common/AppLayout.vue'
 import { roleMenus } from '@/config/roleMenus.js'
-import { useCircularStockStore } from '@/stores/hq/circularStock/circularStock.js'
+import { useCircularStockSaleStore } from '@/stores/hq/circularStock/circularStockSale.js'
 import { circularBuyerApi } from '@/api/hq/circularBuyer.js'
 
 const route = useRoute()
 const router = useRouter()
-const circularStockStore = useCircularStockStore()
+const circularStockStore = useCircularStockSaleStore()
 
 const hqMenus = roleMenus.hq
 const circularStockMenus = roleMenus.hq.find((menu) => menu.label === '순환 재고 관리')?.children ?? []
